--- a/modules/apps/change-tracking/change-tracking-rest-api/src/main/java/com/liferay/change/tracking/rest/resource/v1_0/EntryResource.java
+++ b/modules/apps/change-tracking/change-tracking-rest-api/src/main/java/com/liferay/change/tracking/rest/resource/v1_0/EntryResource.java
@@ -22,6 +22,8 @@ import com.liferay.portal.vulcan.accept.language.AcceptLanguage;
 import com.liferay.portal.vulcan.pagination.Page;
 import com.liferay.portal.vulcan.pagination.Pagination;
 
+import java.util.List;
+
 import javax.annotation.Generated;
 
 import org.osgi.annotation.versioning.ProviderType;
@@ -42,7 +44,7 @@ public interface EntryResource {
 			Long collectionId, String[] changeTypesFilter,
 			String[] classNameIdsFilter, Boolean collision,
 			String[] groupIdsFilter, Integer status, String[] userIdsFilter,
-			Pagination pagination, Sort[] sorts)
+			Pagination pagination, List<Sort> sorts)
 		throws Exception;
 
 	public Entry getEntry(Long entryId) throws Exception;