--- a/modules/apps/change-tracking/change-tracking-rest-impl/src/main/java/com/liferay/change/tracking/rest/internal/resource/v1_0/BaseProcessResourceImpl.java
+++ b/modules/apps/change-tracking/change-tracking-rest-impl/src/main/java/com/liferay/change/tracking/rest/internal/resource/v1_0/BaseProcessResourceImpl.java
@@ -42,13 +42,13 @@ import javax.servlet.http.HttpServletResponse;
 
 import javax.validation.constraints.NotNull;
 
+import javax.ws.rs.BeanParam;
 import javax.ws.rs.DefaultValue;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.Context;
 import javax.ws.rs.core.UriInfo;
 
 /**
@@ -84,7 +84,8 @@ public abstract class BaseProcessResourceImpl implements ProcessResource {
 				com.liferay.change.tracking.rest.constant.v1_0.ProcessType
 					processType,
 			@Parameter(hidden = true) @QueryParam("userId") Long userId,
-			@Context Pagination pagination, @Context Sort[] sorts)
+			@BeanParam Pagination pagination,
+			@Parameter(hidden = true) @QueryParam("sort") List<Sort> sorts)
 		throws Exception {
 
 		return Page.of(Collections.emptyList());