--- a/modules/apps/change-tracking/change-tracking-rest-impl/src/main/java/com/liferay/change/tracking/rest/internal/resource/v1_0/OpenAPIResourceImpl.java
+++ b/modules/apps/change-tracking/change-tracking-rest-impl/src/main/java/com/liferay/change/tracking/rest/internal/resource/v1_0/OpenAPIResourceImpl.java
@@ -25,18 +25,12 @@ import java.util.Set;
 
 import javax.annotation.Generated;
 
-import javax.servlet.ServletConfig;
-
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
-import javax.ws.rs.core.Application;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import javax.ws.rs.core.UriInfo;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -59,18 +53,11 @@ public class OpenAPIResourceImpl {
 	@GET
 	@Path("/openapi.{type:json|yaml}")
 	@Produces({MediaType.APPLICATION_JSON, "application/yaml"})
-	public Response getOpenAPI(
-			@Context HttpHeaders httpHeaders, @Context UriInfo uriInfo,
-			@PathParam("type") String type)
+	public Response getOpenAPI(@PathParam("type") String type)
 		throws Exception {
 
-		return _openAPIResource.getOpenAPI(
-			_application, httpHeaders, _resourceClasses, _servletConfig, type,
-			uriInfo);
+		return _openAPIResource.getOpenAPI(_resourceClasses, type);
 	}
-
-	@Context
-	private Application _application;
 
 	@Reference
 	private OpenAPIResource _openAPIResource;
@@ -92,8 +79,5 @@ public class OpenAPIResourceImpl {
 			add(OpenAPIResourceImpl.class);
 		}
 	};
-
-	@Context
-	private ServletConfig _servletConfig;
 
 }