--- a/modules/apps/data-engine/data-engine-rest-impl/src/main/java/com/liferay/data/engine/rest/internal/resource/v1_0/DataDefinitionResourceImpl.java
+++ b/modules/apps/data-engine/data-engine-rest-impl/src/main/java/com/liferay/data/engine/rest/internal/resource/v1_0/DataDefinitionResourceImpl.java
@@ -46,7 +46,6 @@ import com.liferay.portal.kernel.service.ResourceLocalService;
 import com.liferay.portal.kernel.service.ResourcePermissionLocalService;
 import com.liferay.portal.kernel.service.RoleLocalService;
 import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.odata.entity.EntityModel;
@@ -127,7 +126,8 @@ public class DataDefinitionResourceImpl
 
 	@Override
 	public Page<DataDefinition> getSiteDataDefinitionsPage(
-			Long siteId, String keywords, Pagination pagination, Sort[] sorts)
+			Long siteId, String keywords, Pagination pagination,
+			List<Sort> sorts)
 		throws Exception {
 
 		if (pagination.getPageSize() > 250) {
@@ -137,12 +137,11 @@ public class DataDefinitionResourceImpl
 					"page-size-is-greater-than-x", 250));
 		}
 
-		if (ArrayUtil.isEmpty(sorts)) {
-			sorts = new Sort[] {
+		if (sorts.isEmpty()) {
+			sorts.add(
 				new Sort(
 					Field.getSortableFieldName(Field.MODIFIED_DATE),
-					Sort.STRING_TYPE, true)
-			};
+					Sort.STRING_TYPE, true));
 		}
 
 		return SearchUtil.search(