--- a/modules/apps/headless/headless-common-spi/src/main/resources/com/liferay/headless/common/spi/resource/packageinfo
+++ b/modules/apps/headless/headless-common-spi/src/main/resources/com/liferay/headless/common/spi/resource/packageinfo
@@ -1,1 +1,1 @@
-version 1.0.0
+version 2.0.0