--- a/modules/apps/portal-vulcan/portal-vulcan-api/src/main/resources/com/liferay/portal/vulcan/pagination/packageinfo
+++ b/modules/apps/portal-vulcan/portal-vulcan-api/src/main/resources/com/liferay/portal/vulcan/pagination/packageinfo
@@ -1,1 +1,1 @@
-version 1.0.0
+version 1.1.0