@brianchandotcom baseline is AVE: 37120.6ms, STD: 306.4ms
This pull is AVE: 36913.6ms, STD: 293.3ms

Differences are within noise range, so this pull can be assumed not affecting startup performance.

Somehow this is not generating the same performance benefit as the previous one. I noticed all previous ones are "scope = ServiceScope.PROTOTYPE", and these openapi ones are not.

This may cause the JAX-RS whiteboard code to behave differently.