@liferay-continuous-integration 2019-07-11
CI is automatically triggering &quot;ci:test:sf&quot; and &quot;ci:test:relevant&quot; for this pull to run Source Formatter and relevant tests.

Comment &quot;ci:test&quot; to run the full PR Tester for this pull.