@liferay-continuous-integration 2019-07-11
<h1>Build started.</h1><p>Jenkins is currently running tests.</p><h4>Base Branch:</h4><p>Branch Name: <a href="https://github.com/liferay/liferay-portal/tree/master">master</a></p><h4>Job Summary:</h4><p>Job Link: <a href="https://test-1-20.liferay.com/job/test-portal-source-format/1890/">test-portal-source-format</a></p><p><b>For more details click <a href="https://test-1-20.liferay.com//userContent/jobs/test-portal-source-format/builds/1890/jenkins-report.html">here</a>.</b></p>