@liferay-continuous-integration 2019-07-11
<html><h3>:x: ci:test:sf - 0 out of 1 jobs passed in 48 seconds 34 ms</h3><details><summary>Click here for more details.</summary><h4>Base Branch:</h4><p>Branch Name: <a href="https://github.com/liferay/liferay-portal/tree/master">master</a><br/>Branch GIT ID: <a href="https://github.com/liferay/liferay-portal/commit/43be24b0ac99c6e3a6d30fcb485e4bcb988e7ad2">43be24b0ac99c6e3a6d30fcb485e4bcb988e7ad2</a></p><h4>Sender Branch:</h4><p>Branch Name: <a href="https://github.com/shuyangzhou/liferay-portal/tree/LPS-97971">LPS-97971</a><br/>Branch GIT ID: <a href="https://github.com/shuyangzhou/liferay-portal/commit/1d01fc8a3fb66c1a10871adf2c84e7713a62aa9f">1d01fc8a3fb66c1a10871adf2c84e7713a62aa9f</a></p>0 out of 1jobs PASSED<div><h4>1 Failed Jobs:</h4><ul><li><a href="https://test-1-20.liferay.com/job/test-portal-source-format/1890/"><strike><strong>test-portal-source-format</strong></strike></a></li></ul></div><h5>For more details click <a href="https://test-1-20.liferay.com/userContent/jobs/test-portal-source-format/builds/1890/jenkins-report.html">here</a>.</h5><pre><code>	at org.apache.tools.ant.Task.perform(Task.java:348)
	at org.apache.tools.ant.taskdefs.Sequential.execute(Sequential.java:68)
	at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
	at sun.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:62)
	at sun.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
	at java.lang.reflect.Method.invoke(Method.java:498)
	at org.apache.tools.ant.dispatch.DispatchUtils.execute(DispatchUtils.java:106)
	at org.apache.tools.ant.Task.perform(Task.java:348)
	at net.sf.antcontrib.logic.TryCatchTask.execute(TryCatchTask.java:207)
	at org.apache.tools.ant.UnknownElement.execute(UnknownElement.java:292)
	at sun.reflect.GeneratedMethodAccessor4.invoke(Unknown Source)
	at sun.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
	at java.lang.reflect.Method.invoke(Method.java:498)
	at org.apache.tools.ant.dispatch.DispatchUtils.execute(DispatchUtils.java:106)
	at org.apache.tools.ant.Task.perform(Task.java:348)
	at org.apache.tools.ant.Target.execute(Target.java:435)
	at org.apache.tools.ant.Target.performTasks(Target.java:456)
	at org.apache.tools.ant.Project.executeSortedTargets(Project.java:1393)
	at org.apache.tools.ant.Project.executeTarget(Project.java:1364)
	at org.apache.tools.ant.helper.DefaultExecutor.executeTargets(DefaultExecutor.java:41)
	at org.apache.tools.ant.Project.executeTargets(Project.java:1248)
	at org.apache.tools.ant.Main.runBuild(Main.java:851)
	at org.apache.tools.ant.Main.startAnt(Main.java:235)
	at org.apache.tools.ant.launch.Launcher.run(Launcher.java:280)
	at org.apache.tools.ant.launch.Launcher.main(Launcher.java:109)

BUILD FAILED
/opt/dev/projects/github/liferay-jenkins-ee/commands/build-common.xml:10877: The following error occurred while executing this line:
/opt/dev/projects/github/liferay-jenkins-ee/commands/build-test-portal-source-format.xml:18: Sourced file: inline evaluation of: `` 			 				import com.liferay.jenkins.results.parser.AntUtil; 				import com.lifer . . . '' : TargetError : at Line: 298 : in file: inline evaluation of: `` 			 				import com.liferay.jenkins.results.parser.AntUtil; 				import com.lifer . . . '' : throw e ; 

Target exception: java.lang.RuntimeException: Unable to create local branch temp-master-private1562885149505 at 8ef663032572870f43650fb53daec8e3295f97e1
fatal: Not a valid branch point: '8ef663032572870f43650fb53daec8e3295f97e1'.</code></pre></details></html>