@brianchandotcom 2019-07-11
Hey @nhpatt 

We speculate that @Context only affects FIELD levels, but not method level performance.

From Shuyang:

```
baseline is AVE: 37120.6ms, STD: 306.4ms
This pull is AVE: 36913.6ms, STD: 293.3ms

Differences are within noise range, so this pull can be assumed not affecting startup performance.

Somehow this is not generating the same performance benefit as the previous one. I noticed all previous ones are "scope = ServiceScope.PROTOTYPE", and these openapi ones are not.

This may cause the JAX-RS whiteboard code to behave differently.
```

So let's take a smaller approach, send me a pull to just remove field level @Context. Then, we'll test it with @Context parsing disabled, and with it on, and we should get similar results. If so, we're good to go.

I'd rather not change the API from Sort[] to List<Sort> if it doesn't buy us anything. Sorry about that!

Thx.